Rather than building a kernel.instrumented that everyone forgets to install as kernel, add a WITH_TESLA option that builds an instrumented kernel.  If WITH_TESLA is set it now implies CONFIGARGS=-t.
This change includes partial support for building a nominally
instrumentable kernel without tesla, but it currently includes a
horriable hack that tesla.manifest must always be created.  At
this point this should (assuming it works) be good enough for
benchmarking the impact of tesla on compliation, but we're
probably generalizing from too few examples to nail down the
right solution.

[git-p4: depot-paths = "//depot/projects/ctsrd/tesla/src/": change = 231209]