First cut at an experimental DTrace audit provider, 'dtaudit'.  The provider creates DTrace probes for all known audit events (i.e., those listed in /etc/security/audit_event and pushed into the kernel by auditd when it starts).  When a probe is enabled by a script, additional records may be created that are not otherwise requested by either the trail mechanism or active audit pipes.  These will then be submitted back to DTrace once committed at the end of a system call, in the audit worker thread, where their contents can be inspected by the running script.  The first argument is a struct audit_record pointer -- the in-kernel representation of an audit record before it has been translated to BSM.  The provider will also provide a BSM representation via the second and third arguments (pointer and length).
Currently, user-submitted records are not exposed to DTrace, but we
could easily add support for that -- although they have only BSM, not
an audit_record representation.

In its current form, the provider has one probe per event.  We could
also add 'class' probes allowing collections of events to be
instrumented based on /etc/security/audit_class, to reduce the number
of probes that needs to be named in a script.

Numerous XXX's exist in the source code, which will need sorting out.

Systems must have 'auditd_enable="YES"' in their /etc/rc.conf, and
the audit daemon started, for probes to become visible.