Add a field 'p_uuid' to each process structure, initialised to a fresh UUID when a process is first forked, and before it first returns to userspace.  From that point onward it will be constant.
Teach the audit framework to capture the subject process's UUID when
creating an audit record.

When a new process is forked, capture its UUID via audit, although
currently that information is not exposed via an audit token.  It is
available to the DTrace audit provider.

Currently, random UUIDs based on the host ID are used -- we may want
to move to another model in the future to avoid the need to acquire an
additional global mutex during process fork.