Add a "v_uuid" field to struct vnode.
Initialise as a nil UUID in getnewvnode(9).

Provide a utility function, vn_uuid_from_data(9), that filesystems
can use to populate a vnode's UUID from a filesystem-specific
namespace.  We would ideally incorporate a UUID identifying the
namespace (e.g., the host UUID) but are not yet ready to do that.

For now, all vnode UUIDs are nil.