Add support for LLVM's `llc` tool.
When we use an LLVM IR workflow for individual object files, we need the
llc tool to compile bitcode/IR files into object files. This commit teaches
the build system about the llc tool and includes it in our support for
external toolchains.

This commit does not add an automatic suffix rule for converting .bco/.llo
files into .o files, as that would introduce ambiguity about the correct
path to take from a .c file to a .o. Also, the kernel build (currently the
only place where we will be compiling to objects and then using a
conventional linker) requires extra care and attention, e.g.,
-code-model=kernel.