Rather than creating UUIDs for devfs devices (and their symlinks) in devfs, add a UUID field to 'struct cdev', which describes a character device. This ensures that we will always have access to a suitable UUID when performing device operations (e.g., via device fileops), but also reflects the reality that devfs names are really just pointers to a common set of underlying objects regardless of the specific path (e.g., which devfs mount) is used to get there.  We now properly assign UUIDs for driver-created symlinks for devices, which do have their own cdevs.
There are some gaps in this approach, not least that some devfs entries
are not for devices: subdirectory vnodes and user-created symlinks will
still have nil UUIDs.  We will likely want to introduce UUIDs in another
intermediate layer in devfs: the devfs_dirent structure, as is done for
MAC labels.  Add a comment in the devfs vnode instantiation code to
suggest this path.