Allow suppression of --no-fatal-warnings.
When building with an external toolchain whose linker doesn't support the
--no-fatal-warnings flag, we need a mechanism to suppress it. This mechanism
takes the form of a new make variable, LD_NO_FATAL_WARNS. If it is not set,
it defaults to --Wl,--no-fatal-warnings as now. However, it can be defined
in src.conf, at the command line, etc., to be the empty string.

Sponsored by:   DARPA, AFRL