Introduce (and utilise) the notion that a zero value in a message ID represents an invalid message ID, as the message ID generator starts counting at 1 for each CPU.
Rely on zero filling of the mbuf pkthdr to detect invalid message IDs
so as to avoid including them on message receipt to suppress
notification of messages with non-local origins (which therefore do
not have message IDs).

In the future, we may want to take some more explicit action to assign
message IDs to non-local packets -- e.g., as relates to taggig network
packets.