Increase sophistication of instrumentation option.
Use the options framework from bsd.opts.mk to expose a build option
called MK_INSTRUMENT_BINARIES instead of the more ad-hoc approach we had
been taken of using an environment variable / make argument. This leads
to a more reliable mechanism for disabling binary instrumentation in
particular binaries' makefiles, though it still requires annotation of
each such "special" binary's individual makefile: we still can't disable
the option in an entire subtree.

There could be an argument for putting this option in src.opts.mk
instead of bsd.opts.mk, but we have opted for the latter for two
reasons:

1. we will eventually want to build lots of things with instrumentation,
   so putting the option in bsd.opts.mk will expose it to some extra
   consumers for free (although it's an opt-in option) and
2. it doesn't introduce a dependency from bsd.prog.mk on src.opts.mk,
   which is probably against the rules.