Fix use of C files in LDADD.
When linking instrumented LLVM bitcode together with LDADD, we
expect that this variable's contents are directly-linkable things like
libraries (or, in a pinch, object files). Therefore, full CFLAGS may not
be available at the moment of linking, making this an unsuitable time to
compile C files. Fix this singular use of CFLAGS by explicitly teaching
the build system about the object file that is generated.