Don't include i386 EFI parts twice for amd64.
It looks like the amd64 version of libbfd has been including
efi-app-ia32.c twice in SRCS. This hasn't caused a problem when linking
OBJS, as including the same .o file twice in OBJS doesn't cause a linker
failure. However, this does cause a problem for llvm-link, so it broke
our instrumented-IR builds.