Remove redundant file from libsvn_client SRCS.
For some reason, the file log.c was found twice in libsvn_client's SRCS.
This behaviour tends to be masked by normal compilation workflows, but
when using LLVM IR with llvm-link this is a build error. Fix the issue
for the sake of both our instrumented IR builds and general tidiness.