Make the prototype run. Still issues to resolve.
The main problem is that DTrace currently loads fbt arbitrarily
and there seems to be no easy way to make it not load. That results
on different modules causing problems when it comes to probe numbers
on the host/guest, resulting in a mismatch. This is manifested in
a possible page fault, depending on the modules present on the host
or guest.