Use tail(name) for target rather than name.
As in the upstream version of config(8), use the tail() of a source
filename (e.g., `locore.S`) rather than the full filename (e.g.,
`amd64/amd64/locore.S`) when constructing the corresponding target
filename. This is a bit limiting, as we can't have multiple targets with
the same filename, but that's no more limiting than we experience
upstream.

This may cause issues when building instrumented kernels, but for the
moment the priority needs to be building uninstrumented kernels --- we
can fix instrumentation later if need be.