Redesign how we deal with probes in the guest.
Before, we enabled probes by simply creating an action and bypassing
all of the builtin DTrace mechanisms to maintain state and options.
This commit changes it so that we instead add a new DIF instruction
that fires a hypercall and we establish "virtual" state in DTrace that
is allocated only once. When it is freed, all probes are disabled
with it instead of done one by one.