OpenSSL support for TLS offload.
This patch adds support to our in-tree OpenSSL to support
TLS offload in the transmit direction only.  All apps linking
to the library using normal SSL apis will be accelerated in
the transmit direction, as long as kernel TLS support is
present, and they are using a cipher supported by the kernel.

This patch also adds a new SSL_sendfile() call.

Randall Stewart(rrs@) wrote the initial version of this support,
including the SSL_sendfile() api.

This version is based heavily on Boris Pismenny <borisp@mellanox.com>'s
implementation for Linux kTLS in OpenSSL 1.1, with a backport
to OpenSSL 1.0, support for CBC, and the addition of rrs' wrappers
for sendfile.

Sponsored by: Netflix