Improve DTrace program analysis API.
Require mod/ref analysis consumers to provide policy in the form of a
callback function that says whether or not an action's modifications and
references are OK in light of accumulated mod/ref actions. This also
requires making the API look a bit more "real" in dtrace.h, with a
dedicated part of the file, more #define's and more comments.