Make DLOG a default-off option for now.
Since `dlogd` doesn't build with the new OpenSSL parts from upstream,
disable it by default for now. We can restore it to a default-enabled
option once we've fixed the (API? ABI?) OpenSSL issues, which will
hopefully be trivial for someone with a bit more `dlogd` context than
I have at the moment.