Change message boundary to be message ending in '\n'
This removes the assumption that all messages are JSON data, and that they will
have a message of "}" iff the JSON record is complete.

Newlines seem more intuitive to me, since buffers often flush on newlines. For
performance reasons, it only checks the end of each message for the newline.