Initial (partial) description of CTIR.
CTIR stands for "Concurrent Tracing Intermediate Representation" and
is built on the premise that shared, mutable state can be encoded
as communication between processes. This should then be much more
suitable for statically detecting numerous errors in D scripts.

CTIR is also designed to be extensible and to support creation of
an policy describing how computational resources should be used and
where certain results should be computed.