Still a work in progress.
Currently there seems to be some issue with getting tags to the kernel via
uio. Seems like the first iov is not the one I expect it to be. Still to be
debugged.

This commit fixes the issue of tag propagation from guest kernel (virtio
frontend) to the host userspace (virtio backend).