Initial work on negotiation for probe IDs.
This commit adds the following two functions:
 - dtrace_priv_probeid_adjust
 - dtrace_find_probe

The former is exposed to virtio-dtrace in order to make it possible
to adjust the dtrace_probes array with new probe IDs. We currently
assume the following things:

(i)  The host does not have _more_ probes than the guest. This is an
     assumption that needs to be addressed in the future, but it
     greatly simplifies the code for now as we don't have to
     recalculate the probe IDs on the host in order to have unr work.

(ii) The probe 4-tuples imply that the probes will have the same
     semantics (place of firing in the kernel, same CTF type
     information, etc.).

The latter (dtrace_find_probe) simply takes a probe 4-tuple and finds
the corresponding probe in the current dtrace_probes array. Must be
called with cpu_lock and dtrace_lock held.