Code to generate an ELF file describing a dtrace_prog_t.
This commit implements an ELF generator for a DTrace program.

We assume that translators are applied before the linking stage, so we
don't put any kind of translators inside ELF. This is something that has
to be looked at in the future in order to check these assumptions.

The ELF file that's generated is a 32-bit, little-endian file. It enables
extended section indexing in order to support a split of each action, ECB,
DIFO, variable table, integer table, string table, statement and program
into its own section that references other sections where ever necessary.

In order to create an ELF description of a dtrace_prog_t, the user ought to
call dt_elf_create(dtrace_prog_t *dt_prog, int endian), specifying the
program which has to be converted and the endianness of the resulting ELF
file.