Initial code to parse the ELF file.
We still can not produce a proper working DTrace program from this
as we are missing the following things:

(1) Providers in the dtrace handle
(2) Options from the ELF file (not parsed yet).

and possibly some other state that has yet to be identified. However,
the ELF file is produced _and_ parsed (at least I currently believe)
correctly.