Work in progress.
Add a few NULL checks after mallocs. There is still a problem though:

when the string table is not in the ELF file, neither is any other table.
Moreover:

foo {
  abcdef = 1;
}

does not produce a strtab in the ELF file (it has no data), but:

foo {
  abcdefg = 1;
}

does. Any variable name of size 7 will end up in the string table
and things seem to be fine with that. However:

foo {
  trace(tid);
}

also produces a string table in the ELF file with "tid" inside it.
This is yet to be debugged. Other things, such as running the program
that was taken from the ELF file can be run, but in the case of missing
tables, it will produce wrong results.