Some bugfixes and some TODOs.
We can now correctly parse the options, but we still have to deal with
missing data on the target end. Currently, we are missing the aggregation
ID and the aggregation name. They need to be added (conditionally) for
each statement.

Moreover, added a NOTE that describes the operational model of relinking
on the target machine so that we can get away from doing things in-kernel
where type information is not only not available, but the CTF API is not
implemented.