Add a new string table to the ELF file for identifier names.
This commit creates a new string table in the ELF file that contains
identifier names (e.g. aggregation names) and adds aggdata into each
statement if necessary. This ensures that the target machine trying
to finalise the linking stage of the ELF file does not segfault, and
moreover we have access to both aggregation IDs and their names.