This commit resolves a number of issues.
Firstly, there was an issue where a situation such as:

usetx %r1, sym
usetx %r2, sym
add %r3, %r3, %r2
uload [%r3], %r3
add %r4, %r4, %r1
uload [%r4], %r4

would infer the wrong usetx for each of the uloads. Namely, the first
usetx would be tied to the first uload, and the second usetx to the
second uload. This commit fixes it and checks the information flow
graph to ensure we are dealing with the correct usetx and make sure
that we start relocating from top-down, so that if we encounter a
setx that used to be a usetx relocation, and has been relocated, we
simply skip it and move to the one that wasn't.

Secondly, there was an issue where we would infer new variable types
and as a result change the size of said variable, its CTF identifier
and so on. This commit makes sure that the information gets updated
and that we don't run into the case where we drop variables if we
happen to end up with a calculation of the wrong size where chunksize
ends up being smaller than the expected load size.

Finally, the kernel was missing checks for PUSHTR_G and PUSHTR_H, and
this commit adds those, fixing another part of the chunksize calculation.