Some cleanup with existing HyperTrace files.
This commit cleans up some of the code in pci_virtio_dtrace.c, removing
parts of the code that were under `#if 0`. Additionally, unloading vtdtr
now cleans up all the process queues, freeing all the events and freeing
the queues themselves. Moreover, it adds a new type of event in vtdtr, an
ELF event which represents a path to the ELF file that was just created
containing a new DTrace program. Userspace can actually enqueue this event
(unlike any other event) and notify other processes of it.