Work in progress with dtdaemon.
Now we have a listening and writing thread that is responsible for
putting ELF files on the system and getting them out through a new
kernel module `dttransport`.

dttransport accepts 1 process as a subscriber to DTrace ELF related
events and passes around the ELF file from virtio to userspace and
back.