Still work in progress on dttransport and dtdaemon.
This version implements a fair bit of functionality required to pass
the ELF files back and forth, but there is still a missing part to them.
Namely, checking the SHA256 checksum on every assembly of the ELF file.
This also means that the ELF file can not be parsed by DTrace right now
because there is a checksum slapped on top of the file. This will be
addressed in the next commit.