dtdaemon, in combination with bhyve and dttransport is now able to correctly transmit the file to the guest from a dtrace(1) command, apply the relocations in the guest and transfer the file back to the host, notifying all dtrace(1) instances about the file.
There is still a TODO, though. We need to make sure that the file that
is being sent to the DTrace instance is correctly attributed (in the
sense that when the file arrives, the dtrace(1) instance knows how to
check if the file is meant for it or not). Secondly, we need to fill
in the VM names, which means translating hostnames and FreeBSD versions
to VM names. We then need to extract all the actions in those probes and
add them to the DIFO list so that they can end up in the DOF correctly.