dtdaemon now has a lockfile associated with it in order to address starting multiple instances of it.
Moreover, a number of bugs were fixed and a waschanged() check was removed
because DTrace currently does not support any kind of reinstrumentation
and we are generating random names, so this check alone is not necessary.

It is also a flawed check in this case, as sometimes the file changed time
will change unrelated to DTrace.