Beginning of a virtualized DTrace Framework and some bugfixes.
Each DTrace probe has now become a thing called a _vprobe_. This is essentially
the same as a DTrace probe, however it is further scoped by a virtual DTrace
instance identifier. Specifically, in this case it is the VM ID or 0 (host).

Furthermore, this commit adds vproviders. vproviders are essentially just names
for providers that exist in subordinate DTrace instance (VM), but not on the host.
They do not have any provider operations associated with them, nor can they be
created or unregistered by themselves. They do not own any probes. Instead, the
design choice is to have probes associated with a particular instance (VM) and
have the providers only "exist" as a virtual provider of a given vprobe.

The ioctls to create and delete vprobes is yet to be written.

Finally, this commit introduces the beginning of an implementation for augmented
instrumentation. This is essentially adding ECBs onto the existing DTrace state
for a given consumer. It has not yet been tested, but will be used to dynamically
add instrumentation to VMs and enabling vprobes on the host as certain virtual
machines appear or become responsive (if we asked to instrument them).