Bugfixes & making sensible hypertrace assertions.
There is still a problem where even if hypercalls are enabled, we don't
seem to be getting data from the probes on the host. Moreover, the code
that does a BUFSNAP and processes the buffer doesn't actually get the
probe information for the correct vmid. Instead, it gets the host probe
information, reporting the wrong 4-tuple.