dtdaemon: partial support for different messages.
So far we only supported sending ELF files back and forth. For some
things, we might need to directly control a user process in the guest,
such as killing a dtrace consumer when we are done tracing. This commit
has no functional change other than the plumbing necessary in dtdaemon
to support different messages by adding a message header.