dtdaemon/dtrace/libdtrace: major restructure based on kqueue.
This is still work in progress, but the idea is to remove all of the
manually managed threads for sockets with a central loop that dispatches
events to workers, as it is a simpler design for dealing with
many-to-many relationship of consumers/forwarders -- especially since we
need multiple different kinds of messages instead of just sending around
ELF files.

This also has the nice side-effect that everything stops writing to
/var/ddtrace except for dtdaemon, meaning that all of the state can be
managed in a well-structured and consistent way. It also means that
there is now a protocol to talk to dtdaemon with, rather than an ad-hoc
"put your file in this directory please".