dt_typefile: use the dt_module interface.
So far we have been using a custom interface, but dt_module provides a
good amount of functionality we need (and more for in the future!), so
this commit simply ports the functionality we already had in the API to
use dt_module. This shouldn't affect any other code, but gives us the
ability to have a little bit more functionality for free.