dt_error: handle error reporting correctly.
Previously, we have manually created a file in /var/ddtrace/outbound to
report an error to the host. However, with the dtdaemon changes, this is
no longer the way to do it. This commit simply changes the error
reporting to send the "error" program to dtdaemon and back to the host,
rather than making a file on the filesystem.