dtrace: implement type comparison and subtyping.
Type comparison is relatively straightforward, we mostly just follow C
typing rules. This is necessary because we can no longer compare types
with just checking their ctfids as we are doing so across many
typefiles.

Subtyping is a little bit more complex -- but not too much! The main
difference is that C does not have any kind of subtyping relation, so we
can control the semantics of it. The idea is that in HyperTrace, one
might need different versions of the operating system (or even a
different operating system) to have compatibility between the data being
stored in some way. Instead of relying on manually written translators,
we can define a subtyping relation which gives us the "right one" of the
two types we are comparing. For integers, this is the larger one as it
can represent both accurately. For structures, we require them to be the
same up to a point, but then if one of them extends the structure a bit
further, we will simply return the smaller struct and use that.

We are currently using the (currently rather simplistic) subtyping
relation only for type-checking arguments of a given probe when they
span typefiles, but in theory we could use it in a bit of a more useful
way later.