dtrace: support printf() in D scripts
Prior to this commit, we were just turning printfs into trace() instructions.
This meant that if one wrote a script such as:

```
dtrace -E -n 'foo:syscall::write:entry { printf("hello world: %d\n", arg0); }'
```

the compiler would transform it into:

```
dtrace -E -n 'foo:syscall::write:entry { trace(arg0); }'
```

completely dropping the format buffer.

With this commit, we now preserve the format buffer, but there are a few
subtleties. We aren't actually able to check that the arguments we pass into
printf are of the correct type. This is because the host doesn't have any
type information about the guest, so we just have to assume it will be correct.
The worst case scenario is that the user simply gets runtime failures and
error messages.