dtrace: Initial work on tagging with VM handles.
Before this commit, we couldn't actually identify if memory belongs on the
guest or on the host. This commit starts the work in the direction that not
only allows us to identify if the data is on the guest or host, but it lets us
identify on *which* guest it is.

It is expected that this will have a runtime overhead both in memory footprint
and in CPU cycles, but it is unlikely to be impactful at all compared to the
other parts of HyperTrace. This is something that should be measured, if only
to say that in fact it has very little impact on the overall performance of
the system. One interest point might be how local, non-HyperTrace scripts
behave with this change, though.