dtrace: when inferring the IFG, first compute the basic block and use it later.
Previously, we have relied on a bit of a dirty hack to implement CFG traversal
when computing the information flow graph. This resulted in a bit of a buggy
implementation as indicated by a number of tests in the DTrace test suite.

This commit makes it so that we first compute the corresponding basic block of
a given node and start traversal from from that basic block when computing the
IFG. By doing so, we ensure that we don't end up in a situation where we are
trying to infer a data definition for a node from a node that's not in the
current traversal of the CFG.

We now have 11 broken and 397 failures in the test suite.