dtrace: link the rtype more correctly than before.
Certain actions have a different rtype, even though their DIFO might be
the same. We shouldn't base the rtype just on the stuff that was inferred
in the DIFO, but also on the kind of action. This commit implements a partial
solution to this, as we still need to address aggregations.

Currently we have 3 broken tests and 265 failed tests, as opposed to
12 broken and 282 failed before this patch.