dtrace: work in progress on type-checking st instructions.
So far, we haven't really been handling stb/sth/stw/stx instructions because
I didn't really know when they are generated. After running the DTrace test
suite and examining failures, I now have a better idea of when they are actually
generated and how we might want to handle them.

This commit is an initial take on handling these instructions, but doesn't yet
run as expected because I hadn't implemented a way to infer which variable is
being modified yet. That is to say, dt_get_vardef_from_node() is not yet
implemented and simply returns NULL.