dtrace: infer the IFG from the first instruction instead of last.
This makes it possible for us to also infer variable definitions so that
we can handle st* instructions in typing. This commit also refactors a number
of things, one of which is picking out variables from the DIFO in
dt_populate_varlist() instead of trying to infer it from the store instructions.

This has the effect that we can now begin handling variables declared as a part
of the script, rather than a variable that is defined through DIF loads and
stores.