dtrace: handle variables in the local D data model.
For any variable that is declared ahead of time, e.g. in a script such as

```
struct foo {
  ...
};

struct foo x;

BEGIN {
  x.bar = ...;
}
```

are not a part of the guest's CTF information. We need to process them on
the host and never actually generate the relocations. This commit does exactly
that.