dtrace: infer variables using both the table and our own algorithm.
In some cases, such as the `if` statement in D, DTrace will generate a few
"intermediate" variables that don't actually end up in the variable table.
These are mostly local variables using ldls/stls, and they aren't always
stored to at all, they just get loaded. This means that we need to look at
load instructions as well in order to add variables to the variable table.
Our type-checker will handle all inconsistencies that may arise from this
later on (and hopefully report the error better).