dtrace: initial handling of subroutine semantics based on the subroutine.
DTrace subroutines are kind of weird -- they have a `rd` that they store things
to, but not all of them actually store anything there. We need to handle this
in the IFG correctly so that we don't end up with the wrong type for one of the
definitions (DIF_TYPE_NONE in this case). This commit provides an initial fix
for this problem, and it further refactors typechecking of DTrace subroutines
into its own function in order to cut down on the size of dt_infer_type().