dtrace: handle removing basic blocks from bb_path.
Previously, we haven't really removed any basic blocks from bb_path, which
resulted in type-checking errors because instructions like `flushts` and
`popts` could be skipped due to an unrealistic IFG.

This commit partially fixes this issue by attempting to remove basic blocks
that are no longer used from bb_path. This was sort of happening when we
were using the recursive version of this function (though not correctly),
but was never really ported to the unrolled version. What we currently do is
roughly along the lines of adding every basic block to bb_path as we process
it, and remove it and all of its parents when we either clobber whatever it is
we are looking at (register, variable, stack) or when we reach the leaf of the
CFG. It still doesn't fully work as intended, but it is a start.