dtrace: handle variables declared in the D program itself.
When the user writes a script such as:

```
struct {
  int x;
} y;

BEGIN
{
  y.x = 1;
}
```

the variable `y` ends up in the variable table with a type that the DIF type
inference can never really infer, as this type information is lost once the
D program is "discarded" in favor of the assembled DTrace action format.

I currently think that the only way this can ever be a problem is if the
variable size can't be inferred in the DIF type inference algorithm, so the
kernel gets the wrong variable size and as a result doesn't give us expected
results at runtime. I'm not sure if it can ever end up as a problem where we
need to know this type information in the DIF type inference algorithm itself,
however we will resolve that once it happens.

This commit simply stores the type of a variable generated by D, and if it's
in a D CTF container, it will set a flag to restore the type of the variable
not matter what the DIF typing algorithm thinks about it.