dtrace: handle DIFO return types for trace() and non-trace() actions.
DTrace has a bit of a weird way of encoding trace() actions. Any trace()
action will be encoded as a DTRACEACT_DIFEXPR, which is the same way one
would encode DIFOs that don't actually need to print anything. Unfortunately,
this means that if we don't specially handle trace() actions, we won't be
able to print whatever needs to be printed, and if we let everything be
printed, then a simple expression such as `i++` will print a number every
time it happens.

This commit handles trace expressions in a way that is a little bit messy, but
it works. We store in every action whether or not it needs to have a `void`
return type, and if so, we will patch it up in our linking stage. Otherwise,
we will put in the return type as the one we inferred for the return
register of the DIFO.