dtrace: make it possible to use C and D libdtrace-builtin CTF containers.
This commit adds two new typefiles -- "C" and "D" with the goal of being able
to find various types that aren't really available in the kernel. The side
effect of this is that we can't really call dtrace_update() anymore because
these modules then go missing. Instead, we simply rely on the modules that were
loaded at the time libdtrace was loaded, rather than trying to dynamically
load new modules when we're opening our typefiles. If there's a module that
typefiles can detect, but isn't a libdtrace module, we will simply emit a
warning and skip it.