dtdaemon: -C and -c is confusing, switch it to -O and -m.
dtdaemon used -C for Control machine and -c for Client machines. However,
this is confusing, as even I kept making mistakes in starting up the daemon.
Instead of keeping this confusion around, move the flags to something easier
to remember.

'-O' is what Control used to be and now stands for 'Overlord'.
'-m' is what Client used to be and now stands for 'Minion'

Update the error messages to reflect that and improve their content.