dtraced: Create an init script.
The dtraced init script now supports the following:

dtraced_enable="YES"
dtraced_type="(overlord|minion)"

in rc.conf.

The former specifies whether or not dtraced should be enabled (defaults to NO)
while the latter specifies if dtraced should be run in overlord or minion mode.
There is no default for dtraced_type, and if the user doesn't specify which
mode dtraced should be run in, they will get a warning (and start dtraced
anyway, which will error out).