bhyve: Only enable dthyve if we pass in a '-d' flag.
This is necessary because enabling dthyve opens up a few sockets with dtraced.
These sockets are then expected to be listened on or written to, and if that
doesn't happen, it can lock up a dtraced worker thread. Maybe we shouldn't
have dtraced be that fragile, but for now this simple workaround will do.