dtrace: Ensure that we can trace multiple VMs at the same time.
Without this, we would simply exit on the first VM that responds with having 0
probes. We need *both* fixes, because we don't want to actually have a malicious
VM give us 0 probes back and stop the tracing. Instead, we want to just ignore
it, and report the error unless dtrace is set to quiet mode.