dtrace: Add sanity checks to the probe context.
Note that these printfs are safe to add, because vmid can ever be > 0 (not on
the host) when `dtrace_vprobe()` is called directly, which is only ever done
in the context of a VM calling it. `printf()` itself will never be called in
the probe context on the host, so we won't end up with re-entring probes.