dtrace: Patch up the variable table to handle mismatch in types.
Certain scripts can cause mismatch of type sizes, which then further causes
the verifier to not accept the resulting program. This commit ensures that
the userspace linker patches up the variable table ahead of time to have
matching type sizes on each of the variables, and further handles the BOTTOM
return type for a DIFO by simply using the original one that the host specified.