dtrace: Address a minor race condition in the kernel.
When we are cleaning up probes, we should make sure we first zero all of the
arrays and empty out the hashes. Then we need to sync in order for all of the
CPUs to see the changes as we can't synchronize in dtrace_probe(). Only then can
we free everything.

This is the first commit in a series of commits that will restructure the DTrace
Framework for HyperTrace changes to support race-free disabling of probes and
cleaning up of vprobespaces.