dtrace: Restructure the HyperTrace "provider" into an API.
The HyperTrace provider doesn't really work, as it requires storing multiple
virtual providers, as identified by the probe. The DTrace framework doesn't
really know anything about these vproviders, as they are internal to
HyperTrace. This becomes a problem when the user attempts to enable probes and
needs to match them based on the provider name. It would be extremely
complicated to re-implement the matching algorithm in hypertrace.c, so as such
it seems to be cleaner to simply have a HyperTrace API that is exposed to a
particular tracing framework (in this case, DTrace) that will handle all of
these oprations. Presently, it is pretty messy, but will be cleaned up.

This implies that the DTrace Framework still needs to store these "vproviders"
and there is some duplication -- as well as a possible memory exhaustion attack
available to untrusted guests that needs to be resolved in some way.