dtrace: New command line flag '-r'. Allows running root programs.
The 'root' programs here refers to programs that don't have a srcident. Namely,
those that have been created out of nowhere. This is needed because we don't
want guests to be running untrusted programs from other guests -- instead we
want dtrace(1) to always verify the identifiers that were passed in against the
srcident from the incoming program. The lack '-r' flag simply ensures that we
don't run anything if we don't pass in any identifiers.