dtrace, dtraced: Add configurable sanitizers. Fix a bug.
This commit adds two new /etc/src.conf variables:
 WITH_DTRACE_DEBUG_ASAN
 WITH_DTRACE_DEBUG_TSAN

adding support for ASAN + UBSAN and TSAN. It fails the build if both are set, as
they are incompatible with each other. There is currently a bug where we can't
run make makeman, which needs fixing.

This commit also  fixes a race condition between freeing of a job and accessing
it in another thread. Furthermore, it calls waitpid() if there are no
identifiers for the spawned process and we expect it to exit immediately.