dtraced: Handle state in a more integrated and cleaner way.
This commit moves the logic of starting and stopping threads owned by the state
to init_state() and destroy_state(). In destroy_state(), it also changes the way
thread killing and joining is handled. We don't need to be catching a bunch of
errors and reporting them, as pthread_kill() will only ever error out on an
invalid thread or signal being sent. As we are never sending invalid signals, we
only really have a case where the thread is already dead, and because we are
exiting, we don't really care. We also don't really care if pthread_join()
fails, as we will simply exit the program after the state has been destroyed.