dtraced: Close file descriptors in a race-free way.
This commit adds another thread to dtraced which is responsible for cleaning
up (closing) open file descriptors. This is necessary because prior to this
commit, we don't know if a file descriptor is currently being referenced by
another thread, causing potentially undefined behaviour when close() is called
on it, and another thread calls send() or recv().

The solution implemented in this commit is using simple reference counting and
a separate thread which wakes up every 5 seconds to close unused file
descriptors.